refactor: simplify imports with index in Main
