+++ src/main/config/app.ts
@@ -1,6 +1,5 @@
 import express, { type Express } from 'express'
-import setupMiddlewares from './middlewares'
-import setupRoutes from './routes'
+import { setupMiddlewares, setupRoutes } from '@/main/config'
 
 export const setupApp = async(): Promise<Express> => {
   const app = express()
