+++ src/main/config/index.ts
@@ -1,0 +1,4 @@
+export * from './app'
+export * from './env'
+export * from './middlewares'
+export * from './routes'
