+++ src/main/config/routes.ts
@@ -2,7 +2,7 @@ import { type Express, Router } from 'express'
 import { readdirSync } from 'fs'
 import { join } from 'path'
 
-export default (app: Express): void => {
+export const setupRoutes = (app: Express): void => {
   const router = Router()
   app.use('/api', router)
   readdirSync(join(__dirname, '../routes')).map(async file => {
