+++ src/main/factories/controllers/authentication/index.ts
@@ -1,0 +1,4 @@
+export * from './sign-in-controller-factory'
+export * from './sign-in-validation-factory'
+export * from './sign-up-controller-factory'
+export * from './sign-up-validation-factory'
