+++ src/main/factories/controllers/authentication/sign-in-validation-factory.ts
@@ -1,0 +1,12 @@
+import { ValidationComposite, RequiredFieldValidation, EmailValidation } from '@/validation/validators'
+import { type Validation } from '@/presentation/protocols'
+import { EmailValidatorAdapter } from '@/infra/validators'
+
+export const makeSignInValidation = (): ValidationComposite => {
+  const validations: Validation[] = []
+  for (const field of ['email', 'password']) {
+    validations.push(new RequiredFieldValidation(field))
+  }
+  validations.push(new EmailValidation('email', new EmailValidatorAdapter()))
+  return new ValidationComposite(validations)
+}
