--- src/main/factories/controllers/sign-in-validation-factory.ts
@@ -1,14 +1,0 @@
-import { ValidationComposite } from '@/validation/validators/validation-composite'
-import { type Validation } from '@/presentation/protocols/validation'
-import { RequiredFieldValidation } from '@/validation/validators/required-field-validation'
-import { EmailValidation } from '@/validation/validators/email-validation'
-import { EmailValidatorAdapter } from '@/infra/validators/email-validator-adapter'
-
-export const makeSignInValidation = (): ValidationComposite => {
-  const validations: Validation[] = []
-  for (const field of ['email', 'password']) {
-    validations.push(new RequiredFieldValidation(field))
-  }
-  validations.push(new EmailValidation('email', new EmailValidatorAdapter()))
-  return new ValidationComposite(validations)
-}
