--- src/main/factories/controllers/sign-up-validation-factory.ts
@@ -1,21 +1,0 @@
-import { type Validation } from '@/presentation/protocols/validation'
-import { EmailValidatorAdapter } from '@/infra/validators/email-validator-adapter'
-import { CompareFieldsValidation } from '@/validation/validators/compare-fields-validation'
-import { EmailValidation } from '@/validation/validators/email-validation'
-import { RequiredFieldValidation } from '@/validation/validators/required-field-validation'
-import { ValidationComposite } from '@/validation/validators/validation-composite'
-import { UsernameValidation } from '@/validation/validators/username-validation'
-import { UsernameValidatorAdapter } from '@/infra/validators/username-validator-adapter'
-import { PasswordValidation } from '@/validation/validators/password-validation'
-
-export const makeSignUpValidation = (): ValidationComposite => {
-  const validations: Validation[] = []
-  for (const field of ['username', 'email', 'password', 'passwordConfirmation']) {
-    validations.push(new RequiredFieldValidation(field))
-  }
-  validations.push(new UsernameValidation('username', new UsernameValidatorAdapter()))
-  validations.push(new EmailValidation('email', new EmailValidatorAdapter()))
-  validations.push(new PasswordValidation('password'))
-  validations.push(new CompareFieldsValidation('password', 'passwordConfirmation'))
-  return new ValidationComposite(validations)
-}
