+++ src/main/middlewares/index.ts
@@ -1,0 +1,3 @@
+export * from './body-parser'
+export * from './content-type'
+export * from './cors'
