+++ src/main/routes/authentication-route.ts
@@ -1,7 +1,6 @@
 import { type Router } from 'express'
-import { adaptRoute } from '../adapters/express-route-adapter'
-import { makeSignUpController } from '../factories/controllers/sign-up-controller-factory'
-import { makeSignInController } from '../factories/controllers/sign-in-controller-factory'
+import { adaptRoute } from '@/main/adapters'
+import { makeSignUpController, makeSignInController } from '@/main/factories/controllers/authentication'
 
 export default (router: Router): void => {
   router.post('/signup', adaptRoute(makeSignUpController()))
