+++ src/main/server.ts
@@ -1,6 +1,6 @@
 import 'module-alias/register'
-import { MongoHelper } from '@/infra/db/mongodb/mongo-helper'
-import env from './config/env'
+import { MongoHelper } from '@/infra/db/mongodb'
+import { env } from '@/main/config'
 
 MongoHelper.connect(env.mongoUrl)
   .then(async() => {
